+++ CHANGELOG.md
@@ -44,6 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1012](https://github.com/alleslabs/celatone-frontend/pull/1012) Move modules, balances query to LCD
+
 ### Bug fixes
 
 - [#1004](https://github.com/alleslabs/celatone-frontend/pull/1004) Fix empty string moniker